code/machine learning/brain/utils/image_processing.py
import requests
from io import BytesIO
import random
from PIL import Image
import os
import time

def preprocess_image(image_url):
    """
    Download and preprocess an image from a URL
    """
    try:
        # Check if the URL is a local file path
        if image_url.startswith('http'):
            # Download image from URL
            response = requests.get(image_url)
            image = Image.open(BytesIO(response.content))
        else:
            # Remove leading slash if present
            if image_url.startswith('/'):
                image_url = image_url[1:]

            # Handle relative paths
            if not os.path.isabs(image_url):
                # Assume the path is relative to the backend directory
                base_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(__file__))))
                image_path = os.path.join(base_dir, image_url)
            else:
                image_path = image_url

            # Open the local file
            image = Image.open(image_path)

        # Convert to RGB if necessary
        if image.mode != 'RGB':
            image = image.convert('RGB')

        # Resize to standard size
        image = image.resize((224, 224))

        # Simulate processing delay
        time.sleep(random.uniform(0.1, 0.5))

        # Return the image (in a real system, we would convert to numpy array)
        return image
    except Exception as e:
        raise Exception(f"Error preprocessing image: {str(e)}")

def enhance_image(image):
    """
    Simulate enhancing image quality
    """
    # In a real system, we would apply contrast enhancement, etc.
    # Here we just return the original image
    return image

def segment_tumor(image):
    """
    Simulate segmenting tumor region from the image
    """
    # In a real system, we would apply thresholding, find contours, etc.
    # Here we just return a simulated tumor contour
    width, height = image.size
    center_x = width // 2
    center_y = height // 2
    radius = min(width, height) // 4

    # Simulate a circular tumor contour
    tumor_contour = {
        'center': (center_x, center_y),
        'radius': radius
    }

    return tumor_contour

def calculate_tumor_dimensions(tumor_contour, pixel_spacing=(0.1, 0.1)):
    """
    Calculate tumor dimensions in centimeters
    """
    if tumor_contour is None:
        return None

    # In a real system, we would calculate dimensions from the contour
    # Here we just return simulated dimensions
    radius_cm = tumor_contour['radius'] * pixel_spacing[0]
    diameter_cm = radius_cm * 2
    area_cm2 = 3.14159 * radius_cm * radius_cm

    return {
        'width': diameter_cm,
        'height': diameter_cm,
        'area': area_cm2
    }

def extract_features(image):
    """
    Simulate extracting features from the image
    """
    # In a real system, we would calculate statistics, texture features, etc.
    # Here we just return simulated features
    features = {
        'mean': random.uniform(0.4, 0.6),
        'std': random.uniform(0.1, 0.3),
        'min': random.uniform(0.0, 0.2),
        'max': random.uniform(0.8, 1.0),
        'entropy': random.uniform(3.0, 5.0)
    }

    return features