code/machine learning/brain/requirements.txt
flask
flask-cors
numpy
pillow
python-dotenv
gunicorn